+++ src/mako/http/exceptions/HttpException.php
@@ -23,6 +23,13 @@ class HttpException extends RuntimeException
 	protected $defaultMessage;
 
 	/**
+	 * Additional data.
+	 *
+	 * @var array
+	 */
+	protected $data = [];
+
+	/**
 	 * Constructor.
 	 *
 	 * @param int             $code     Exception code
@@ -32,5 +39,28 @@ class HttpException extends RuntimeException
 	public function __construct(int $code, string $message = '', ?Throwable $previous = null)
 	{
 		parent::__construct($message ?: (string) $this->defaultMessage, $code, $previous);
+	}
+
+	/**
+	 * Sets additional data.
+	 *
+	 * @param  array $data Additional data
+	 * @return $this
+	 */
+	public function setData(array $data)
+	{
+		$this->data = $data;
+
+		return $this;
+	}
+
+	/**
+	 * Returns additional data.
+	 *
+	 * @return array
+	 */
+	public function getData(): array
+	{
+		return $this->data;
 	}
 }