+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### New
 
 * The default production error views now have "dark mode" support.
+* Added `setData` and `getData` methods to the `HttpException` class.
+* The `HttpException` data array is available in the production error templates as `$data`.
 
 --------------------------------------------------------